+++ places/admin.py
@@ -1,6 +1,25 @@
 from django.contrib import admin
 
-from .models import Place, PlaceType
+from .models import Place, PlaceType, ResidenceType
+
+
+@admin.register(ResidenceType)
+class ResidenceTypeAdmin(admin.ModelAdmin):
+    list_display = (
+        "code",
+        "name",
+        "sort_order",
+        "is_active",
+        "is_system",
+    )
+    search_fields = (
+        "code",
+        "name",
+    )
+    list_filter = (
+        "is_active",
+        "is_system",
+    )
 
 
 admin.site.register(PlaceType)